@@ -29,7 +29,7 @@ import (
 	oconfig "github.com/IBM/ibm-block-csi-operator/pkg/config"
 	oversion "github.com/IBM/ibm-block-csi-operator/version"
 	"github.com/go-logr/logr"
-	"github.com/presslabs/controller-util/syncer"
+	"github.com/presslabs/controller-util/pkg/syncer"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
